Throwing an API_Exception if the uploaded file fails to copy in APIv3

+++ api/v3/Attachment.php
@@ -146,7 +146,9 @@ function civicrm_api3_attachment_create($params) {
   elseif (is_string($moveFile)) {
     // CRM-17432 Do not use rename() since it will break file permissions.
     // Also avoid move_uplaoded_file() because the API can use options.move-file.
-    copy($moveFile, $path);
+    if (!copy($moveFile, $path)) {
+      throw new API_Exception("Cannot copy uploaded file ".$moveFile." to ".$path);
+    }
     unlink($moveFile);
   }
